requirements.txt
coverage==4.3.4
decorator==4.0.11
doctest2==0.1
docutils==0.13.1
nose==1.3.7
pathlib2==2.2.1
poyo==0.4.0
proc==0.11
pytest==3.0.6
PyYAML==3.12
requests==2.12.5
unittest2==1.1.0
virtualenv==15.0.3
codecov==2.0.5
